requirements.txt
numpy>=1.23
openai>=1.0.0
scikit-learn==1.2.2
python-dotenv==0.21.0
requests>=2.28.0
six==1.16.0
fpdf2>=2.7.0
azure-ai-inference>=0.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
tqdm>=4.65.0
langchain>=0.0.267
langchain-openai>=0.0.2
pydantic>=2.0.0
pytest>=7.0.0
colorama>=0.4.4